content: generateScopedTrackChronologyLinks
Content in generateTrackChronologyLinks is a stub.
